+++ include/ui/include.hpp
@@ -5,6 +5,9 @@
 #include "utils.hpp"
 #include "setup.hpp"
 #include "constants.hpp"
+#include "syntax.hpp"
+#include "textures.hpp"
+#include "tooltip.hpp"
 
 namespace nwo5::ui::prelude {
     namespace ui = nwo5::ui;
@@ -18,4 +21,6 @@ namespace nwo5::ui::prelude {
     using ui::Font;
 
     using ui::Alignment;
+
+    using ui::Tooltip, ui::TooltipInfo;
 }