+++ include/ui/nodes.hpp
@@ -50,6 +50,7 @@ namespace nwo5::ui {
     SILLY_API_DLL Setup<geode::AxisLayout> column(Alignment pAlignment, std::optional<Alignment> pCrossAlignment = std::nullopt);
 
     SILLY_API_DLL geode::AxisLayoutOptions* options();
+    SILLY_API_DLL geode::AnchorLayoutOptions* anchorOptions();
 
     /// no auto scale, grows axis, no cross axis
     /// @deprecated use the other layout functions