+++ include/ui/textures.hpp
@@ -1,0 +1,131 @@
+#pragma once
+
+namespace nwo5::ui {
+    namespace frame {
+        static constexpr auto GREEN_ARROW = "GJ_arrow_01_001.png";
+        static constexpr auto BLUE_ARROW = "GJ_arrow_02_001.png";
+        static constexpr auto PINK_ARROW = "GJ_arrow_03_001.png";
+        static constexpr auto SIMPLE_ARROW = "navArrowBtn_001.png";
+
+        static constexpr auto UP_ARROW = "edit_upBtn_001.png";
+        static constexpr auto DOWN_ARROW = "edit_downBtn_001.png";
+        static constexpr auto LEFT_ARROW = "edit_leftBtn_001.png";
+        static constexpr auto RIGHT_ARROW = "edit_rightBtn_001.png";
+        
+        static constexpr auto HORIZONTAL_ARROWS = "edit_flipXBtn_001.png";
+        static constexpr auto VERTICAL_ARROWS = "edit_flipYBtn_001.png";
+
+        static constexpr auto CW_ARROW = "edit_cwBtn_001.png";
+        static constexpr auto CCW_ARROW = "edit_ccwBtn_001.png";
+
+        static constexpr auto RED_X = "GJ_deleteIcon_001.png";
+
+        static constexpr auto LOCK = "GJ_lock_001.png";
+        static constexpr auto OPEN_LOCK = "GJ_lock_open_001.png";
+        static constexpr auto GRAY_LOCK = "GJ_lockGray_001.png";
+
+        static constexpr auto STAR = "GJ_starsIcon_001.png";
+        static constexpr auto GRAY_STAR = "GJ_starsIcon_gray_001.png";
+
+        static constexpr auto PLUS = "edit_addCBtn_001.png";
+        static constexpr auto MINUS = "edit_delCBtn_001.png";
+
+        static constexpr auto LIKE = "GJ_likesIcon_001.png";
+        static constexpr auto DISLIKE = "GJ_dislikesIcon_001.png";
+
+        static constexpr auto FOLDER = "folderIcon_001.png";
+
+        static constexpr auto HAMMERS = "GJ_hammerIcon_001.png";
+        
+        static constexpr auto HELP = "GJ_infoIcon_001.png";
+
+        static constexpr auto NOTE = "GJ_noteIcon_001.png";
+        static constexpr auto MUSIC = "GJ_noteIcon_001.png";
+        static constexpr auto NOTE_ALT = "GJ_musicIcon_001.png";
+        static constexpr auto MUSIC_ALT = "GJ_musicIcon_001.png";
+
+        static constexpr auto FILTER = "GJ_filterIcon_001.png";
+
+        static constexpr auto SORT = "GJ_sortIcon_001.png";
+
+        static constexpr auto FIND = "edit_findBtn_001.png";
+
+        static constexpr auto CHECK = "GJ_completesIcon_001.png";
+        
+        static constexpr auto DOWNLOAD = "GJ_downloadsIcon_001.png";
+
+        static constexpr auto TRASH = "edit_delBtn_001.png";
+
+        static constexpr auto EXCLAMATION_MARK = "exMark_001.png";
+        static constexpr auto EXCLAMATION_POINT = "exMark_001.png";
+  
+        static constexpr auto STAR_BUTTON = "GJ_starBtn_001.png";
+
+        static constexpr auto BACK_BUTTON = "GJ_backBtn_001.png";
+
+        static constexpr auto LIKE_BUTTON = "GJ_likeBtn_001.png";
+        static constexpr auto DISLIKE_BUTTON = "GJ_dislikeBtn_001.png";
+
+        static constexpr auto INFO_BUTTON = "GJ_infoBtn_001.png";
+
+        static constexpr auto MENU_BUTTON = "GJ_menuBtn_001.png";
+
+        static constexpr auto PLUS_BUTTON = "GJ_plusBtn_001.png";
+
+        static constexpr auto UNDO_BUTTON = "GJ_undoBtn_001.png";
+        static constexpr auto REDO_BUTTON = "GJ_redoBtn_001.png";
+
+        static constexpr auto BLUE_PAUSE_BUTTON = "GJ_pauseBtn_001.png";
+        static constexpr auto PAUSE_BUTTON = "GJ_pauseBtn_001.png";
+        static constexpr auto YELLOW_PAUSE_BUTTON = "GJ_pauseEditorBtn_001.png";
+        static constexpr auto EDITOR_PAUSE_BUTTON = "GJ_pauseEditorBtn_001.png";
+
+        static constexpr auto TRASH_BUTTON = "GJ_trashBtn_001.png";
+        
+        static constexpr auto REFRESH_BUTTON = "GJ_updateBtn_001.png";
+        static constexpr auto WHITE_REFRESH_BUTTON = "GJ_updateBtn_001.png";
+        static constexpr auto REFRESH_BUTTON_ALT = "GJ_getSongInfoBtn_001.png";
+        static constexpr auto WHITE_REFRESH_BUTTON_ALT = "GJ_getSongInfoBtn_001.png";
+        static constexpr auto UPDATE_BUTTON = "GJ_updateBtn_001.png";
+        static constexpr auto BLUE_REFRESH_BUTTON = "GJ_replayBtn_001.png";
+        static constexpr auto REPLAY_BUTTON = "GJ_replayBtn_001.png";
+
+        static constexpr auto GEAR_BUTTON = "GJ_optionsBtn_001.png";
+        static constexpr auto COG_BUTTON = "GJ_optionsBtn_001.png";
+        static constexpr auto OPTIONS_BUTTON = "GJ_optionsBtn_001.png";
+
+        static constexpr auto HELP_BUTTON = "GJ_helpBtn_001.png";
+
+        static constexpr auto AUDIO_BUTTON = "GJ_fxOnBtn_001.png";
+        static constexpr auto AUDIO_OFF_BUTTON = "GJ_fxOffBtn_001.png";
+        static constexpr auto MUSIC_BUTTON = "GJ_musicOnBtn_001.png";
+        static constexpr auto MUSIC_OFF_BUTTON = "GJ_musicOffBtn_001.png";
+
+        static constexpr auto YELLOW_X_BUTTON = "GJ_closeBtn_001.png";
+        static constexpr auto CLOSE_BUTTON = "GJ_closeBtn_001.png";
+        static constexpr auto RED_X_BUTTON = "GJ_deleteBtn_001.png";
+        static constexpr auto DELETE_BUTTON = "GJ_deleteBtn_001.png";
+        static constexpr auto RED_X_BUTTON_ALT = "GJ_deleteSongBtn_001.png";
+        static constexpr auto DELETE_BUTTON_ALT = "GJ_deleteSongBtn_001.png";
+    }
+    namespace sprite {
+        static constexpr auto SQUARE = "square02_001.png";
+        static constexpr auto BACKGROUND = "GJ_square01.png";
+        static constexpr auto BROWN_BACKGROUND = "GJ_square01.png";
+        static constexpr auto BROWN_SQUARE = "GJ_square01.png";
+        static constexpr auto DEFAULT_BACKGROUND = "GJ_square01.png";
+        static constexpr auto DEFAULT_SQUARE = "GJ_square01.png";
+        static constexpr auto BLUE_BACKGROUND = "GJ_square02.png";
+        static constexpr auto BLUE_SQUARE = "GJ_square02.png";
+        static constexpr auto GREEN_BACKGROUND = "GJ_square03.png";
+        static constexpr auto GREEN_SQUARE = "GJ_square03.png";
+        static constexpr auto PURPLE_BACKGROUND = "GJ_square04.png";
+        static constexpr auto PURPLE_SQUARE = "GJ_square04.png";
+        static constexpr auto GRAY_BACKGROUND = "GJ_square05.png";
+        static constexpr auto GRAY_SQUARE = "GJ_square05.png";
+        static constexpr auto LIGHT_GRAY_BACKGROUND = "GJ_square06.png";
+        static constexpr auto LIGHT_SQUARE = "GJ_square06.png";
+        static constexpr auto EMPTY_BACKGROUND = "GJ_square07.png";
+        static constexpr auto EMPTY_SQUARE = "GJ_square07.png";
+    }
+}